backend/requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cmake==3.17.1
colorama==0.4.3
config==0.4.2
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
data==0.4
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==2.2.12
django-cors-headers==3.2.1
djangorestframework==3.11.0
dlib==19.19.0
dnspython==1.16.0
drf-yasg==1.17.1
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
funcsigs==1.0.2
gast==0.2.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.4.0
inflection==0.4.0
ipykernel==5.2.0
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itypes==1.2.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mysql-connector-python==8.0.19
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.2.0.34
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.18
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
win-unicode-console==0.5
wrapt==1.11.2
zipp==1.2.0
absl-py==0.9.0
acme==0.31.0
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
Automat==0.6.0
backcall==0.1.0
bleach==3.1.3
blinker==1.4
cachetools==4.0.0
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cloud-init==18.3
cmake==3.17.1
colorama==0.4.3
command-not-found==0.3
config==0.4.2
ConfigArgParse==0.11.0
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cycler==0.10.0
data==0.4
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
Django==2.2.12
django-cors-headers==3.2.1
djangorestframework==3.11.0
dlib==19.19.0
dnspython==1.16.0
drf-yasg==1.17.1
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
funcsigs==1.0.2
future==0.15.2
gast==0.2.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
hibagent==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.4.0
incremental==16.10.1
inflection==0.4.0
ipykernel==5.2.0
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itypes==1.2.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
josepy==1.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
mysql-connector-python==8.0.19
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.4
ndg-httpsclient==0.4.4
netifaces==0.10.4
notebook==6.0.3
numpy==1.18.3
oauthlib==3.1.0
opencv-python==4.2.0.34
packaging==20.3
PAM==0.4.2
pandas==1.0.3
pandocfilters==1.4.2
parsedatetime==2.4
parso==0.6.2
pbr==3.1.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
Pygments==2.6.1
pygobject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
PyMySQL==0.8.0
pyOpenSSL==17.5.0
pyparsing==2.4.6
pyRFC3339==1.0
pyrsistent==0.15.7
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
pywinpty==0.5.7
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==16.0.0
six==1.14.0
sqlparse==0.3.1
ssh-import-id==5.7
systemd-python==234
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
Twisted==17.9.0
typed-ast==1.4.1
ufw==0.35
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.18
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
win-unicode-console==0.5
wrapt==1.11.2
zipp==1.2.0
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
